+++ djpcms/forms/cms.py
@@ -103,9 +103,9 @@ class PageForm(forms.ModelForm):
         2) 
     '''
     site             = forms.ModelChoiceField(queryset = Site.objects.all(), required = False)
-    application_view = forms.LazyChoiceField(choices = siteapp_choices,
-                                             required = False,
-                                             label = _('application view'))
+    application_view = forms.ChoiceField(choices = siteapp_choices,
+                                         required = False,
+                                         label = _('application view'))
     
     class Meta:
         model = Page
@@ -247,7 +247,7 @@ class ContentBlockForm(EditingForm):
     plugin_name     = PluginChoice(label = _('Plugin'),
                                    choices = plugingenerator,
                                    required = False)
-    container_type  = forms.LazyChoiceField(label=_('Container'), choices = wrappergenerator)
+    container_type  = forms.ChoiceField(label=_('Container'), choices = wrappergenerator)
     view_permission = forms.ModelMultipleChoiceField(queryset = Group.objects.all(), required = False)
     layout = FormLayout(Fieldset('plugin_name','container_type','title','view_permission'),
                         Columns(('for_not_authenticated',),('requires_login',), css_class=inlineLabels3))