+++ djpcms/utils/html/__init__.py
@@ -1,7 +1,6 @@
 from pagination import *
 from base import *
 from grid960 import *
-from autocomplete import *
 
 
 def box(hd = None, bd = None, ft = None, minimize = False, rounded = True,
@@ -29,3 +28,4 @@ def box(hd = None, bd = None, ft = None, minimize = False, rounded = True,
     return loader.render_to_string(['box.html',
                                     'content/box.html',
                                     'djpcms/content/box.html'], c)
+